Currently working Crossover
Implemented crossover without demes and elitism  and prints 3 diagrams out. It is significantly faster than hillclimber which is good news.

It might be a good idea to plot individuals climbing and see if their patterns are correct or not.

However the scatter still doesn't stick to the surface...don't know why.

Next step is to implement with demes and elitism. Maybe play around with mutation rates