Currently working hillclimber with demes
Restructured code to give functions which plot generically. Restructured the hillclimber code to give the same plots with multiple demes...however I need to check why the highest fitness individual seems to go down...Is this why elitism is necessary? But the highest fitness individual should never be replaced...

To be continued...