Working crossover and hillclimber with demes and migration
In the title, however the fittest individual seems to have it's fitness go up and down again, this might have to do with the array being confused and appending things from other demes onto the array for one deme. Need to check

Also need to reimplement the GAs with wheel selection and replacing the entire population with the children of the selected parents instead of just one. This needs to be done for both hillclimber and crossover. This should mean the generation count will be more akin to his on his paper