Currently working all with old GA, fixed fittest going down
Need to re-implement both the GAs for wheel selection. But fixed the problem where the fittestZ array would go down in fitness. This was due to protecting the fittest individual from migration before the crossover and mutation, instead of after. This meant that the protected individual may not have been the actual fittest individual after all the mutation and crossover.